Wrap migration SQL in migration do-block, emit with pgx
This change introduces a build script which wraps the contents of each
SQL file in the `migrations` directory in a DO block which conditionally
applies the migration if it has not yet been applied to the database.

The build script outputs its contents to `migration.sql`, which in turn
is included in the generated SQL which pgx outputs.